Cambios de int16 a int

+++ Expresion/Aritmetica/Negativo.cs
@@ -24,7 +24,7 @@ namespace Pascal_AirMax.Expresion.Aritmetica
 
             if(res_right.getTipo() == Objeto.TipoObjeto.INTEGER)
             {
-                return new Primitivo(Objeto.TipoObjeto.INTEGER, Int16.Parse(res_right.getValor().ToString()) * -1);
+                return new Primitivo(Objeto.TipoObjeto.INTEGER, int.Parse(res_right.getValor().ToString()) * -1);
             }else if(res_right.getTipo() == Objeto.TipoObjeto.REAL)
             {
                 return new Primitivo(Objeto.TipoObjeto.REAL, Double.Parse(res_right.getValor().ToString()) * -1);
